deployment/requirements.txt
pandas==2.2.2
streamlit==1.44.0
seaborn==0.13.2
matplotlib==3.9.4
scipy==1.13.1
numpy==1.26.4
plotly==5.24.1
pillow==11.2.1
tensorflow==2.18.0
